feat: add delete functionality + delete button appearence

+++ src/components/Card.js
@@ -11,6 +11,8 @@ export class Card {
     this._name = data.name;
     this._image = data.link;
     this._id = data._id;
+    this._ownerData = data.owner;
+    this._ownerId = this._ownerData._id;
     this._userId = userId;
     this._likes = data.likes;
     this._templateSelector = templateSelector;
@@ -23,14 +25,17 @@ export class Card {
     const cardElement = document.querySelector(this._templateSelector).content.querySelector('.cards__item').cloneNode(true);
     return cardElement;
   }
+
 //Установка слушателей: попап карточки, удаление, лайк соответственно
   _setEventListeners() {
     this._cardImage.addEventListener('click', () => {
       this._handleCardClick(this._obj);
     });
-    this._cardDelete.addEventListener('click', () => {
-      this._handleDeleteClick(this._id, this._newCard);
-    });
+    if (this._ownerId == this._userId) {
+      this._cardDelete.addEventListener('click', () => {
+        this._handleDeleteClick(this._id, this._newCard);
+      });
+    }
     this._cardLike.addEventListener('click', () => {
       this._handleLikeClick(this._cardLike);
     });
@@ -45,6 +50,10 @@ export class Card {
     this._cardLike = this._newCard.querySelector('.cards__like');
     this._newCard.querySelector('.cards__like-counter').textContent = this._likes.length;
     this._cardDelete = this._newCard.querySelector('.cards__delete');
+    if (this._ownerId !== this._userId) {
+      this._cardDelete.remove();
+    }
+    console.log(this._likes);
     this._setEventListeners();
     return this._newCard;
   }
